feat: 새 캡슐함 생성하기
